backend/requirements.txt
apistar==0.3.9
Django==2.0.1
json_environ==0.1.1
celery==4.1.0
scrapy==1.5.0
redis==2.10.6
scrapyd==1.2.0
python-scrapyd-api==2.0.1
django-redis==4.8.0
gunicorn==19.7.1
